requirements.txt
opencv-contrib-python>=4.8.0
numpy>=1.24.0
pydantic>=2.0.0

# Object detection
ultralytics>=8.0.0

# Object tracking
norfair>=2.2.0
supervision>=0.17.0

# Machine learning
scikit-learn>=1.3.0

# Data export
pandas>=2.0.0
pyarrow>=14.0.0

# Database (TimescaleDB + pgvector)
psycopg2-binary>=2.9.0
pgvector>=0.2.0

# CLIP and VLM
torch>=2.0.0
torchvision>=0.15.0
ftfy
regex
pillow>=9.0.0

# LLM
langchain
langchain-mistralai
langgraph

# Progress bars
tqdm>=4.65.0
